+++ src/DataType.php
@@ -15,37 +15,37 @@ enum DataType: string
     case Item = 'item';
 
     /**
-     * @throws StructuredFieldError
+     * @throws SyntaxError
      */
-    public function fromRfc9651(Stringable|string $httpValue): StructuredField
+    public function fromRfc9651(Stringable|string $httpValue): OuterList|InnerList|Parameters|Dictionary|Item
     {
         return $this->parse($httpValue, Ietf::Rfc9651);
     }
 
     /**
-     * @throws StructuredFieldError
+     * @throws SyntaxError
      */
-    public function fromRfc8941(Stringable|string $httpValue): StructuredField
+    public function fromRfc8941(Stringable|string $httpValue): OuterList|InnerList|Parameters|Dictionary|Item
     {
         return $this->parse($httpValue, Ietf::Rfc8941);
     }
 
     /**
-     * @throws StructuredFieldError
+     * @throws SyntaxError
      */
-    public function parse(Stringable|string $httpValue, ?Ietf $rfc = null): StructuredField
+    public function parse(Stringable|string $httpValue, ?Ietf $rfc = null): OuterList|InnerList|Parameters|Dictionary|Item
     {
         return match ($this) {
             self::List => OuterList::fromHttpValue($httpValue, $rfc),
-            self::InnerList => InnerList::fromHttpValue($httpValue, $rfc),
-            self::Parameters => Parameters::fromHttpValue($httpValue, $rfc),
             self::Dictionary => Dictionary::fromHttpValue($httpValue, $rfc),
             self::Item => Item::fromHttpValue($httpValue, $rfc),
+            self::InnerList => InnerList::fromHttpValue($httpValue, $rfc),
+            self::Parameters => Parameters::fromHttpValue($httpValue, $rfc),
         };
     }
 
     /**
-     * @throws StructuredFieldError
+     * @throws SyntaxError
      */
     public function toRfc9651(iterable $data): string
     {
@@ -53,7 +53,7 @@ enum DataType: string
     }
 
     /**
-     * @throws StructuredFieldError
+     * @throws SyntaxError
      */
     public function toRfc8941(iterable $data): string
     {
@@ -61,7 +61,7 @@ enum DataType: string
     }
 
     /**
-     * @throws StructuredFieldError
+     * @throws SyntaxError
      */
     public function serialize(iterable $data, ?Ietf $rfc = null): string
     {
@@ -69,16 +69,16 @@ enum DataType: string
     }
 
     /**
-     * @throws StructuredFieldError
+     * @throws SyntaxError
      */
-    public function create(iterable $data): StructuredField
+    public function create(iterable $data): OuterList|InnerList|Parameters|Dictionary|Item
     {
         return match ($this) {
             self::List => OuterList::fromPairs($data),
-            self::InnerList => InnerList::fromPair([...$data]), /* @phpstan-ignore-line */
-            self::Parameters => Parameters::fromPairs($data),
             self::Dictionary => Dictionary::fromPairs($data),
-            self::Item => Item::fromPair([...$data]), /* @phpstan-ignore-line */
+            self::Item => Item::fromPair([...$data]),
+            self::InnerList => InnerList::fromPair([...$data]),
+            self::Parameters => Parameters::fromPairs($data),
         };
     }
 }