+++ CHANGELOG.md
@@ -39,7 +39,7 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 - `Parameters:pair` replaced by `Parameters::getByIndex`
 - `Dictionary:get` replaced by `Dictionary::getByKey`
 - `Dictionary:pair` replaced by `Dictionary::getByIndex`
-
+- `::hasNoMember` and `::hasMembers` methods have been replaced by `isEmpty` and `isNotEmpty` on containers
 
 ## [1.3.0](https://github.com/bakame-php/http-structured-fields/compare/1.2.2...1.3.0) - 2024-01-05